refactor(canvas-render): share the one reversal-count test oracle
The two new path-reversal test files carried byte-identical copies of the
independent per-axis reversal oracle. Moved to test-utils per the repo's
shared-test-helper rule; still never calls production code.
